Samyak/requirements.txt
backports.csv==1.0.7
backports.shutil-get-terminal-size==1.0.0
braintree==3.56.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
crypto==1.4.1
decorator==4.0.10
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.1
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-import-export==1.2.0
django-mathfilters==0.4.0
django-paytm==0.1
enum34==1.1.6
et-xmlfile==1.0.1
firebase-admin==3.0.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.0
httplib2==0.14.0
idna==2.8
instamojo-wrapper==1.1
ipython-genutils==0.1.0
jdcal==1.4.1
msgpack==0.6.2
Naked==0.1.31
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.1.0
prompt-toolkit==1.0.9
protobuf==3.9.2
ptyprocess==0.5.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycrypto==2.6.1
Pygments==2.1.3
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.12.0
stripe==1.82.1
tablib==0.13.0
traitlets==4.3.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
